Refator: 프로젝트 폴더 구조 개선 - commponents 폴더의 세부 컴포넌트를 역할에 맞게 구분하여 정리 - app 라우팅에서 main 페이지를 home으로 변경
Relate to: #27
