+++ src/app/(authorized)/(home)/@search/page.tsx
@@ -1,0 +1,9 @@
+import { SearchInput } from "@/components/Inputs";
+
+export default function SearchPage() {
+  return (
+    <div className="flex flex-col gap-6">
+      <SearchInput />
+    </div>
+  );
+}
