+++ src/app/(authorized)/(home)/layout.tsx
