--- src/app/(authorized)/(main)/@search/page.tsx
@@ -1,9 +1,0 @@
-import SearchBar from "@/components/search/SearchBar";
-
-export default function SearchPage() {
-  return (
-    <div className="flex flex-col gap-6">
-      <SearchBar />
-    </div>
-  );
-}
