+++ src/app/(authorized)/@modal/(.)feed/[id]/page.tsx
@@ -1,5 +1,5 @@
-import { FeedDetailCard } from "@/components/feed/card";
-import { InterceptingModal } from "@/components/modal";
+import { FeedDetailCard } from "@/components/Cards/FeedCard";
+import { InterceptingModal } from "@/components/Modals";
 
 export default function FeedDetailModal() {
   return (
