+++ src/app/(authorized)/@modal/(.)feed/create/page.tsx
@@ -1,5 +1,5 @@
-import { FeedCreateCard } from "@/components/feed/card";
-import { InterceptingModal } from "@/components/modal";
+import { FeedCreateCard } from "@/components/Cards/FeedCard";
+import { InterceptingModal } from "@/components/Modals";
 
 export default function FeedCreateModal() {
   return (
