+++ src/app/(authorized)/feed/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { FeedDetailCard } from "@/components/feed/card";
+import { FeedDetailCard } from "@/components/Cards/FeedCard";
 
 export default function FeedDetailPage() {
   return <FeedDetailCard />;
