--- src/app/error/page.tsx
@@ -1,3 +1,0 @@
-export default function ErrorPage() {
-  return <p>Sorry, something went wrong</p>;
-}
