+++ src/components/Buttons/index.ts
@@ -1,0 +1,3 @@
+export { default as FileUploadButton } from "./FileUploadButton";
+export { default as ModalBackButton } from "./ModalBackButton";
+export { default as SubmitButton } from "./SubmitButton";
