+++ src/components/Cards/FeedCard/index.ts
@@ -1,0 +1,3 @@
+export { default as FeedCreateCard } from "./FeedCreateCard";
+export { default as FeedDetailCard } from "./FeedDetailCard";
+export { default as FeedEditCard } from "./FeedEditCard";
