+++ src/components/Feeds/index.ts
@@ -1,0 +1,4 @@
+export { default as EmptyFeed } from "./EmptyFeed";
+export { default as Feed } from "./Feed";
+export { default as FeedList } from "./FeedList";
+export { default as FeedTopMenu } from "./FeedTopMenu";
