+++ src/components/Forms/index.ts
@@ -1,0 +1,1 @@
+export { default as FeedForm } from "./FeedForm";
