+++ src/components/Modals/index.ts
@@ -1,0 +1,3 @@
+export { default as DeleteCheckModal } from "./DeleteCheckModal";
+export { default as InterceptingModal } from "./InterceptingModal";
+export { default as ModalPage } from "./ModalPage";
