--- src/components/feed/card/index.tsx
@@ -1,3 +1,0 @@
-export { default as FeedDetailCard } from "./FeedDetailCard";
-export { default as FeedCreateCard } from "./FeedCreateCard";
-export { default as FeedEditCard } from "./FeedEditCard";
