--- src/components/feed/list/index.tsx
@@ -1,4 +1,0 @@
-export { default as EmptyFeed } from "./EmptyFeed";
-export { default as Feed } from "./Feed";
-export { default as FeedList } from "./FeedList";
-export { default as TopMenu } from "./TopMenu";
