--- src/components/modal/check/index.tsx
@@ -1,1 +1,0 @@
-export { default as DeleteConfirmModal } from "./DeleteConfirmModal";
