--- src/components/modal/index.tsx
@@ -1,2 +1,0 @@
-export { default as InterceptingModal } from "./InterceptingModal";
-export { default as ModalPage } from "./ModalPage";
